requirements.txt
addict
future
lmdb
numpy
opencv-python
Pillow
pyyaml
requests
scikit-image
scipy
torch
torchvision
tqdm
yapf
lpips
gdown 
timm
tensorboard
pyiqa